Se modificó la plantilla CSV para incluir el tipo de medida, se ajustaron los apartados de reportes e inventario: en reportes ahora se muestra el tipo de medida y en inventario se valida que los productos con unidad no permitan cantidades decimales.

+++ UsaloYa.Services/interfaces/ISaleService.cs
@@ -12,7 +12,7 @@ namespace UsaloYa.Services.interfaces
     {
         Task<int> AddSale(SaleDto sale);
         Task<bool> AddProductsToSale(int saleId, List<SaleDetailsDto> saleDetails);
-        Task<bool> UpdateStock(int productId, int selledItems);
+        Task<bool> UpdateStock(int productId, decimal selledItems);
         Task<bool> UpdateTotalSale(int saleId, decimal totalSale);
         Task<bool> UpdateSaleStatus(int saleId, SaleStatus status);
         Task<bool> UpdateStockAfterStatusChange(int saleId, int companyId, bool isCancelAction);
